Further updates to recent QuickCheck and criterion.  aern-order now compiles on ghc 7.8.4 and recent hackage.

+++ aern-order/src/Numeric/AERN/NumericOrder/Arbitrary.hs
@@ -197,12 +197,12 @@ instance
         [elem] <- gen
         return $ UniformlyOrderedSingleton elem
         where
-        Just gen = arbitraryTupleInAreaRelatedBy area [1] []
+        Just gen = arbitraryTupleInAreaRelatedBy area [1::Int] []
 
 instance
     (ArbitraryOrderedTuple t) 
     => 
-    Arbitrary (UniformlyOrderedPair t) 
+    Arbitrary (UniformlyOrderedPair t)
     where
     arbitrary =
         do
